+++ persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/Community.java
@@ -29,7 +29,7 @@ public class Community implements Serializable {
     private Integer id;
 
     @Column(name = "appVersion", insertable = false, updatable = false)
-    private Integer appVersionId;
+    private byte appVersionId;
 
     @ManyToOne(fetch = EAGER)
     @JoinColumn(name = "appVersion")
